services/coqui-engine/requirements.txt
fastapi
uvicorn
prometheus-client
coqui-tts
torch
pydantic
